requirement.txt
arabic-reshaper==2.1.4
argcomplete==2.0.0
asgiref==3.5.2
asn1crypto==1.5.1
autopep8==2.0.0
awscli==1.27.43
boto3==1.26.43
botocore==1.29.43
certifi==2022.9.24
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
cryptography==38.0.3
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-import-export==3.0.1
docutils==0.16
durationpy==0.5
environ==1.0
et-xmlfile==1.1.0
future==0.18.2
hjson==3.1.0
html5lib==1.1
idna==3.4
jmespath==1.0.1
kappa==0.6.0
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.3.0
placebo==0.9.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tablib==3.2.1
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
troposphere==4.2.0
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
Werkzeug==2.2.2
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zappa==0.56.1
